Separate serialization test code (#66)
* Seperate serialization test

* Add test script for serialization

* Add more serialization test code in script
