--- test/transactionType/chainDataAnchoring.js
@@ -1,73 +1,0 @@
-/*
-    Copyright 2018 The caver-js Authors
-    This file is part of the caver-js library.
- 
-    The caver-js library is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Lesser General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
- 
-    The caver-js library is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-    GNU Lesser General Public License for more details.
- 
-    You should have received a copy of the GNU Lesser General Public License
-    along with the caver-js. If not, see <http://www.gnu.org/licenses/>.
-*/
-
-require('it-each')({ testPerIteration: true })
-const { expect } = require('../extendedChai')
-
-const testRPCURL = require('../testrpc')
-const Caver = require('../../index.js')
-
-let caver
-
-beforeEach(() => {
-  caver = new Caver(testRPCURL)
-})
-
-describe('ServiceChain: Chain data anchoring', () => {
-  const sender_transaction = {
-    type: 'CHAIN_DATA_ANCHORING',
-    from: '0x90B3E9A3770481345A7F17f22f16D020Bccfd33e',
-    to: '0x75c3098Be5E4B63FBAc05838DaAEE378dD48098d',
-    nonce: '0x11',
-    gasPrice: '0x19',
-    gas: '0x3b9ac9ff',
-    value: '0x989680',
-    anchoredData: '0xf8a6a00000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000001a00000000000000000000000000000000000000000000000000000000000000002a00000000000000000000000000000000000000000000000000000000000000003a0000000000000000000000000000000000000000000000000000000000000000405',
-    chainId: '0x1',
-  }
-
-  const expectedRawTransaction = '0x48f901251119843b9ac9ff9475c3098be5e4b63fbac05838daaee378dd48098d839896809490b3e9a3770481345a7f17f22f16d020bccfd33eb8a8f8a6a00000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000001a00000000000000000000000000000000000000000000000000000000000000002a00000000000000000000000000000000000000000000000000000000000000003a0000000000000000000000000000000000000000000000000000000000000000405f844f84225a0578e55342e5d4a1e685e2e332dbe5534fb0c0c19cfe7cf84a1be9461bf876ee99f97529a612ab958bb332316d513ddbb57c11d3f8b1e0f951b05ae791dd225e0'
-  
-  it('CAVERJS-UNIT-SER-004 : Sign transaction', async () => {
-    const privateKey = '0xf8cc7c3813ad23817466b1802ee805ee417001fcce9376ab8728c92dd8ea0a6b'
-
-    caver.klay.accounts.wallet.add(privateKey)
-
-    const { rawTransaction } = await caver.klay.accounts.signTransaction(sender_transaction, privateKey)
-
-    expect(rawTransaction).to.equal(expectedRawTransaction)
-
-  }).timeout(200000)
-
-  it('CAVERJS-UNIT-SER-044: Decode raw transaction', async () => {
-    const txObj = await caver.klay.decodeTransaction(expectedRawTransaction)
-    
-    expect(txObj).not.to.be.undefined
-    expect(txObj.type).to.equals(sender_transaction.type)
-    expect(caver.utils.hexToNumber(txObj.nonce)).to.equals(caver.utils.hexToNumber(sender_transaction.nonce))
-    expect(caver.utils.hexToNumber(txObj.gasPrice)).to.equals(caver.utils.hexToNumber(sender_transaction.gasPrice))
-    expect(caver.utils.hexToNumber(txObj.gas)).to.equals(caver.utils.hexToNumber(sender_transaction.gas))
-    expect(txObj.to).to.equals(sender_transaction.to)
-    expect(txObj.value).to.equals(sender_transaction.value)
-    expect(txObj.from).to.equals(sender_transaction.from)
-    expect(txObj.anchoredData).to.equals(sender_transaction.anchoredData)
-    expect(txObj.v).not.to.be.undefined
-    expect(txObj.r).not.to.be.undefined
-    expect(txObj.s).not.to.be.undefined
-  }).timeout(200000)
-})
