+++ test/transactionType/contractExecution.js
@@ -28,51 +28,6 @@ var senderPrvKey
 var senderAddress
 var testAccount
 
-describe('Contract: Contract execution', () => {
-  const sender_transaction = {
-    type: 'SMART_CONTRACT_EXECUTION',
-    from: '0x90B3E9A3770481345A7F17f22f16D020Bccfd33e',
-    to: '0x5e008646fde91fb6eda7b1fdabc7d84649125cf5',
-    nonce: '0xb',
-    data: '0x6353586b00000000000000000000000090b3e9a3770481345a7f17f22f16d020bccfd33e',
-    gas: '0x3b9ac9ff',
-    gasPrice: '0x19',
-    value: '0xa',
-    chainId: '0x1',
-  }
-
-  const expectedRawTransaction = '0x30f89e0b19843b9ac9ff945e008646fde91fb6eda7b1fdabc7d84649125cf50a9490b3e9a3770481345a7f17f22f16d020bccfd33ea46353586b00000000000000000000000090b3e9a3770481345a7f17f22f16d020bccfd33ef845f84325a0028f3080511406a4bab318001e44b5b6fd408ba27f45b376ec2e1949f364b570a0154d385b9addf7e39f1e48f84397cab85a95bb6baae884f6ed6f885d1c48f506'
-  
-  it('CAVERJS-UNIT-SER-006 : Sign transaction', async () => {
-    caver = new Caver(testRPCURL)
-    const privateKey = '0xf8cc7c3813ad23817466b1802ee805ee417001fcce9376ab8728c92dd8ea0a6b'
-
-    caver.klay.accounts.wallet.add(privateKey)
-
-    const { rawTransaction } = await caver.klay.accounts.signTransaction(sender_transaction, privateKey)
-
-    expect(rawTransaction).to.equal(expectedRawTransaction)
-
-  }).timeout(200000)
-
-  it('CAVERJS-UNIT-SER-046: Decode raw transaction', async () => {
-    const txObj = await caver.klay.decodeTransaction(expectedRawTransaction)
-    
-    expect(txObj).not.to.be.undefined
-    expect(txObj.type).to.equals(sender_transaction.type)
-    expect(caver.utils.hexToNumber(txObj.nonce)).to.equals(caver.utils.hexToNumber(sender_transaction.nonce))
-    expect(caver.utils.hexToNumber(txObj.gasPrice)).to.equals(caver.utils.hexToNumber(sender_transaction.gasPrice))
-    expect(caver.utils.hexToNumber(txObj.gas)).to.equals(caver.utils.hexToNumber(sender_transaction.gas))
-    expect(txObj.to).to.equals(sender_transaction.to)
-    expect(caver.utils.hexToNumber(txObj.value)).to.equals(caver.utils.hexToNumber(sender_transaction.value))
-    expect(txObj.from).to.equals(sender_transaction.from)
-    expect(txObj.data).to.equals(sender_transaction.data)
-    expect(txObj.v).not.to.be.undefined
-    expect(txObj.r).not.to.be.undefined
-    expect(txObj.s).not.to.be.undefined
-  }).timeout(200000)
-})
-
 describe('SMART_CONTRACT_EXECUTION transaction', () => {
   var executionObject
   var contractAddress
