+++ test/transactionType/feeDelegatedCancelTransaction.js
@@ -42,62 +42,6 @@ before(() => {
   testAccount = caver.klay.accounts.wallet.add(caver.klay.accounts.create())
 })
 
-describe('Cancel: Fee Delegated Cancel Transaction', () => {
-  const sender_transaction = {
-    type: 'FEE_DELEGATED_CANCEL',
-    from: '0x90B3E9A3770481345A7F17f22f16D020Bccfd33e',
-    nonce: 15,
-    gas: '0x3b9ac9ff',
-    gasPrice: '0x19',
-    chainId: '0x1',
-  }
-  const feePayer = '0x33f524631e573329a550296f595c820d6c65213f'
-  const expectedRawTransaction = '0x39f8bf0f19843b9ac9ff9490b3e9a3770481345a7f17f22f16d020bccfd33ef845f84326a04309d80412c206e604771e2a4b15017f2a0bdfc1f1dd28c9c63dabdc7e8d9ff0a05b98c2c1bb2faae610f791e424444664d6f18d8ed8787bfe66ea6672d974ef349433f524631e573329a550296f595c820d6c65213ff845f84325a062fbaa9cf65a57f61dd680e13f93cacf8ea064211f111cae11222cae6fdfde77a060df7c7b9958d2e10568df83a8e84e7c73b4549075957c344a7848d16fd34915'
-  
-  it('CAVERJS-UNIT-SER-009 : Sign transaction', async () => {
-    const privateKey = '0xf8cc7c3813ad23817466b1802ee805ee417001fcce9376ab8728c92dd8ea0a6b'
-
-    caver.klay.accounts.wallet.add(privateKey)
-
-    const { rawTransaction: senderRawTransaction } = await caver.klay.accounts.signTransaction(sender_transaction, privateKey)
-    const decoded = await caver.klay.decodeTransaction(senderRawTransaction)
-    expect(decoded.feePayer).to.equals('0x')
-    expect(decoded.payerV).to.equals('0x01')
-    expect(decoded.payerR).to.equals('0x')
-    expect(decoded.payerS).to.equals('0x')
-
-    const fee_payer_transaction = {
-      senderRawTransaction: senderRawTransaction,
-      feePayer,
-      chainId: '0x1',
-    }
-
-    const feePayerPrivateKey = '0xb9d5558443585bca6f225b935950e3f6e69f9da8a5809a83f51c3365dff53936'
-    const { rawTransaction } = await caver.klay.accounts.signTransaction(fee_payer_transaction, feePayerPrivateKey)
-
-    expect(rawTransaction).to.equal(expectedRawTransaction)
-
-  }).timeout(200000)
-
-  it('CAVERJS-UNIT-SER-049: Decode raw transaction', async () => {
-    const txObj = await caver.klay.decodeTransaction(expectedRawTransaction)
-
-    expect(txObj).not.to.be.undefined
-    expect(txObj.type).to.equals(sender_transaction.type)
-    expect(caver.utils.hexToNumber(txObj.nonce)).to.equals(caver.utils.hexToNumber(sender_transaction.nonce))
-    expect(caver.utils.hexToNumber(txObj.gasPrice)).to.equals(caver.utils.hexToNumber(sender_transaction.gasPrice))
-    expect(caver.utils.hexToNumber(txObj.gas)).to.equals(caver.utils.hexToNumber(sender_transaction.gas))
-    expect(txObj.from).to.equals(sender_transaction.from)
-    expect(txObj.v).not.to.be.undefined
-    expect(txObj.r).not.to.be.undefined
-    expect(txObj.s).not.to.be.undefined
-    expect(txObj.feePayer).to.equals(feePayer)
-    expect(txObj.payerV).not.to.be.undefined
-    expect(txObj.payerR).not.to.be.undefined
-    expect(txObj.payerS).not.to.be.undefined
-  }).timeout(200000)
-})
-
 describe('FEE_DELEGATED_CANCEL transaction', () => {
   var cancelObject
