+++ test/transactionType/feeDelegatedContractExecution.js
@@ -28,69 +28,6 @@ var senderPrvKey
 var senderAddress
 var testAccount
 
-describe('Contract: Fee Delegated Contract Execution', () => {
-  const sender_transaction = {
-    type: 'FEE_DELEGATED_SMART_CONTRACT_EXECUTION',
-    from: '0x90B3E9A3770481345A7F17f22f16D020Bccfd33e',
-    to: '0x5e008646fde91fb6eda7b1fdabc7d84649125cf5',
-    nonce: 12,
-    value: '0xa',
-    data: '0x6353586b00000000000000000000000090b3e9a3770481345a7f17f22f16d020bccfd33e',
-    gas: '0x3b9ac9ff',
-    gasPrice: '0x0',
-    chainId: '0x1',
-  }
-  const feePayer = '0x33f524631e573329a550296f595c820d6c65213f'
-  const expectedRawTransaction = '0x31f8fa0c80843b9ac9ff945e008646fde91fb6eda7b1fdabc7d84649125cf50a9490b3e9a3770481345a7f17f22f16d020bccfd33ea46353586b00000000000000000000000090b3e9a3770481345a7f17f22f16d020bccfd33ef845f84326a0b85779d9e3a78189925de6cccbadf31183805f3ee7ce04da259551ed4bd1f697a0712fbeb1fcf1c418f6868c3e442cdc223a7d11f70fad240f2870cbfc4463c8249433f524631e573329a550296f595c820d6c65213ff845f84326a0e597315f311abcedeb03c9f0dce0808b70843250b60351fb2171bbef86dd92d8a038dab32a7e686bbfada4e107b7ba507027f07ae143a92296670c85028381602c'
-  
-  it('CAVERJS-UNIT-SER-013 : Sign transaction', async () => {
-    caver = new Caver(testRPCURL)
-    const privateKey = '0xf8cc7c3813ad23817466b1802ee805ee417001fcce9376ab8728c92dd8ea0a6b'
-
-    caver.klay.accounts.wallet.add(privateKey)
-
-    const { rawTransaction: senderRawTransaction } = await caver.klay.accounts.signTransaction(sender_transaction, privateKey)
-    const decoded = await caver.klay.decodeTransaction(senderRawTransaction)
-    expect(decoded.feePayer).to.equals('0x')
-    expect(decoded.payerV).to.equals('0x01')
-    expect(decoded.payerR).to.equals('0x')
-    expect(decoded.payerS).to.equals('0x')
-
-    const fee_payer_transaction = {
-      senderRawTransaction: senderRawTransaction,
-      feePayer,
-      chainId: '0x1',
-    }
-
-    const feePayerPrivateKey = '0xb9d5558443585bca6f225b935950e3f6e69f9da8a5809a83f51c3365dff53936'
-    const { rawTransaction } = await caver.klay.accounts.signTransaction(fee_payer_transaction, feePayerPrivateKey)
-
-    expect(rawTransaction).to.equal(expectedRawTransaction)
-
-  }).timeout(200000)
-
-  it('CAVERJS-UNIT-SER-053: Decode raw transaction', async () => {
-    const txObj = await caver.klay.decodeTransaction(expectedRawTransaction)
-    
-    expect(txObj).not.to.be.undefined
-    expect(txObj.type).to.equals(sender_transaction.type)
-    expect(caver.utils.hexToNumber(txObj.nonce)).to.equals(caver.utils.hexToNumber(sender_transaction.nonce))
-    expect(caver.utils.hexToNumber(txObj.gasPrice)).to.equals(caver.utils.hexToNumber(sender_transaction.gasPrice))
-    expect(caver.utils.hexToNumber(txObj.gas)).to.equals(caver.utils.hexToNumber(sender_transaction.gas))
-    expect(txObj.to).to.equals(sender_transaction.to)
-    expect(caver.utils.hexToNumber(txObj.value)).to.equals(caver.utils.hexToNumber(sender_transaction.value))
-    expect(txObj.from).to.equals(sender_transaction.from)
-    expect(txObj.data).to.equals(sender_transaction.data)
-    expect(txObj.v).not.to.be.undefined
-    expect(txObj.r).not.to.be.undefined
-    expect(txObj.s).not.to.be.undefined
-    expect(txObj.feePayer).to.equals(feePayer)
-    expect(txObj.payerV).not.to.be.undefined
-    expect(txObj.payerR).not.to.be.undefined
-    expect(txObj.payerS).not.to.be.undefined
-  }).timeout(200000)
-})
-
 describe('FEE_DELEGATED_SMART_CONTRACT_EXECUTION transaction', () => {
   var executionObject
   var contractAddress
