+++ test/transactionType/valueTransfer.js
@@ -42,47 +42,6 @@ before(() => {
   testAccount = caver.klay.accounts.wallet.add(caver.klay.accounts.create())
 })
 
-describe('Value transfer: Value Transfer', () => {
-  const sender_transaction = {
-    type: 'VALUE_TRANSFER',
-    from: '0x90B3E9A3770481345A7F17f22f16D020Bccfd33e',
-    to: '0x75c3098Be5E4B63FBAc05838DaAEE378dD48098d',
-    nonce: '0x1',
-    gas: '0x3b9ac9ff',
-    gasPrice: '0x19',
-    value: '0x989680',
-    chainId: '0x1',
-  }
-  const expectedRawTransaction = '0x08f87c0119843b9ac9ff9475c3098be5e4b63fbac05838daaee378dd48098d839896809490b3e9a3770481345a7f17f22f16d020bccfd33ef845f84325a079793f7c580e206b08fd6997df2bbeb89726fa9830c79231adb1b8aba40099f5a07a84e1833cc7ab78300e72bed861ee9b20f52709a8ea43d3d20ffec1c723e1d0'
-
-  it('CAVERJS-UNIT-SER-020 : Sign transaction', async () => {
-    const privateKey = '0xf8cc7c3813ad23817466b1802ee805ee417001fcce9376ab8728c92dd8ea0a6b'
-
-    caver.klay.accounts.wallet.add(privateKey)
-
-    const { rawTransaction } = await caver.klay.accounts.signTransaction(sender_transaction, privateKey)
-
-    expect(rawTransaction).to.equal(expectedRawTransaction)
-
-  }).timeout(200000)
-
-  it('CAVERJS-UNIT-SER-060: Decode raw transaction', async () => {
-    const txObj = await caver.klay.decodeTransaction(expectedRawTransaction)
-    
-    expect(txObj).not.to.be.undefined
-    expect(txObj.type).to.equals(sender_transaction.type)
-    expect(caver.utils.hexToNumber(txObj.nonce)).to.equals(caver.utils.hexToNumber(sender_transaction.nonce))
-    expect(caver.utils.hexToNumber(txObj.gasPrice)).to.equals(caver.utils.hexToNumber(sender_transaction.gasPrice))
-    expect(caver.utils.hexToNumber(txObj.gas)).to.equals(caver.utils.hexToNumber(sender_transaction.gas))
-    expect(txObj.to).to.equals(sender_transaction.to)
-    expect(caver.utils.hexToNumber(txObj.value)).to.equals(caver.utils.hexToNumber(sender_transaction.value))
-    expect(txObj.from).to.equals(sender_transaction.from)
-    expect(txObj.v).not.to.be.undefined
-    expect(txObj.r).not.to.be.undefined
-    expect(txObj.s).not.to.be.undefined
-  }).timeout(200000)
-})
-
 describe('VALUE_TRANSFER transaction', () => {
   var valueTransferObject
